ai/requirements.txt
# Python dependencies for Video Game Recommendation System
# Install with: pip install -r requirements.txt

numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.28.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
